src/multiple_state_monitor/src/requirements.txt
influxdb-client==1.36.0
paho-mqtt==2.1.0

